requirements.txt
PyYAML==6.0.1
Jinja2==3.1.2
Flask==2.3.3
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0